rename some files, debug funtions included in a single file to be easily removed, new endpoint to get single token information
