+++ api/debug.go
@@ -5,28 +5,57 @@ import (
 	"database/sql"
 	"encoding/json"
 	"errors"
+	"math/big"
+	"strconv"
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/vocdoni/census3/census"
 	queries "github.com/vocdoni/census3/db/sqlc"
 	"go.vocdoni.io/dvote/httprouter"
 	api "go.vocdoni.io/dvote/httprouter/apirest"
 	"go.vocdoni.io/dvote/log"
 )
 
-// initHoldersHandlers function registers the endpoints related with the token
-// holders such us get a token holders lists of a token. It uses the given
-// initialized API.
-func (capi *census3API) initHoldersHandlers() {
-	// TODO: Move the createToken endpoint to other api package file.
+// TODO: Only for the MVP, remove it.
+func (capi *census3API) initDebugHandlers() {
 	capi.endpoint.RegisterMethod("/tokens/{address}/holders", "GET",
 		api.MethodAccessTypePublic, capi.getTokenHolders)
 	capi.endpoint.RegisterMethod("/tokens/{address}/holders/count", "GET",
 		api.MethodAccessTypePublic, capi.countHolders)
+	capi.endpoint.RegisterMethod("/census/{censusID}/check/{root}", "POST",
+		api.MethodAccessTypePublic, capi.checkIPFSCensus)
+}
+
+// createDummyStrategy creates the default strategy for a given token. This
+// basic strategy only includes the holders of the given token which have a
+// balance positive balance (holder_balance > 0).
+//
+// TODO: Only for the MVP, remove it.
+func (capi *census3API) createDummyStrategy(tokenID []byte) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	res, err := capi.sqlc.CreateStategy(ctx, "test")
+	if err != nil {
+		return err
+	}
+	strategyID, err := res.LastInsertId()
+	if err != nil {
+		return err
+	}
+	_, err = capi.sqlc.CreateStrategyToken(ctx, queries.CreateStrategyTokenParams{
+		StrategyID: strategyID,
+		TokenID:    tokenID,
+		MinBalance: big.NewInt(0).Bytes(),
+		MethodHash: []byte("test"),
+	})
+	return err
 }
 
 // getTokenHolders handler function gets the token holders states from the
 // database, of the token identified by the contract address provided.
+//
+// TODO: Only for the MVP, remove it.
 func (capi *census3API) getTokenHolders(msg *api.APIdata, ctx *httprouter.HTTPContext) error {
 	ctx2, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
@@ -69,7 +98,8 @@ func (capi *census3API) getTokenHolders(msg *api.APIdata, ctx *httprouter.HTTPCo
 
 // countHolders handler function returns the current number of token holders
 // for the provided token address in the current database.
-// TODO: Only for debug, consider to delete
+//
+// TODO: Only for the MVP, remove it.
 func (capi *census3API) countHolders(msg *api.APIdata, ctx *httprouter.HTTPContext) error {
 	ctx2, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
@@ -99,4 +129,30 @@ func (capi *census3API) countHolders(msg *api.APIdata, ctx *httprouter.HTTPConte
 		return ErrEncodeTokenHolders.Withf("token address: %s", addr)
 	}
 	return ctx.Send(response, api.HTTPstatusOK)
+}
+
+// TODO: Only for the MVP, remove it.
+func (capi *census3API) checkIPFSCensus(msg *api.APIdata, ctx *httprouter.HTTPContext) error {
+	// get strategy and query about the tokens that it includes
+	censusID, err := strconv.Atoi(ctx.URLParam("censusID"))
+	root := ctx.URLParam("root")
+	if err != nil {
+		return ErrMalformedStrategyID
+	}
+
+	internalCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	currentCensus, err := capi.sqlc.CensusByID(internalCtx, int64(censusID))
+	if err != nil {
+		return ErrCantGetStrategy
+	}
+
+	censusDef := census.DefaultCensusDefinition(censusID, 0, nil)
+	censusDef.URI = currentCensus.Uri.String
+	if err := capi.censusDB.Check(censusDef, []byte(root)); err != nil {
+		log.Error(err)
+		return api.APIerror{Code: 5100, HTTPstatus: 500, Err: err}
+	}
+
+	return ctx.Send([]byte("ok"), api.HTTPstatusOK)
 }
